requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.1
alembic-postgresql-enum==1.7.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
bottle==0.13.3
Bottleneck==1.4.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==44.0.2
dataclasses-json==0.6.7
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docx2txt==0.9
email_validator==2.2.0
emoji==2.14.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastcrud==0.15.6
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
gevent==25.4.2
google-ai-generativelanguage==0.6.18
google-api-core==2.24.0
google-api-python-client==2.168.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==0.4.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
identify==2.6.9
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.3
langchain==0.3.24
langchain-community==0.3.19
langchain-core==0.3.56
langchain-google-genai==2.1.4
langchain-google-vertexai==2.0.21
langchain-huggingface==0.1.2
langchain-ollama==0.3.2
langchain-openai==0.3.14
langchain-postgres==0.0.13
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.1
langgraph-checkpoint==2.0.24
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.13
lxml==5.4.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==2.0.2
oauthlib==3.2.2
olefile==0.47
ollama==0.4.8
onnxruntime==1.21.0
openai==1.76.0
opencv-python==4.11.0.86
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pg8000==1.31.2
pgvector==0.3.6
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.51
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.7
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.7.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdf==5.3.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.15
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rq==2.3.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scramp==1.4.5
sentence-transformers==3.4.1
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
twilio==9.5.2
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.35.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
vine==5.1.0
virtualenv==20.29.3
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
